(PC-35111)[API] chore: Create new module serializers. Move offers serialization and schema into this new module
