+++ api/src/pcapi/serializers/thumbnails_serialize.py
