+++ api/tests/core/offers/test_api.py
@@ -38,7 +38,6 @@ from pcapi.core.offers import exceptions
 from pcapi.core.offers import factories
 from pcapi.core.offers import models
 from pcapi.core.offers import repository as offers_repository
-from pcapi.core.offers import schemas as offers_schemas
 from pcapi.core.offers.exceptions import NotUpdateProductOrOffers
 from pcapi.core.offers.exceptions import ProductNotFound
 from pcapi.core.providers.allocine import get_allocine_products_provider
@@ -56,6 +55,7 @@ from pcapi.models.offer_mixin import OfferStatus
 from pcapi.models.offer_mixin import OfferValidationStatus
 from pcapi.models.offer_mixin import OfferValidationType
 from pcapi.notifications.push import testing as push_testing
+from pcapi.serializers import offers_serialize
 from pcapi.utils.date import timespan_str_to_numrange
 from pcapi.utils.human_ids import humanize
 
@@ -1171,7 +1171,7 @@ class CreateMediationV2Test:
 class CreateDraftOfferTest:
     def test_create_draft_offer_from_scratch(self):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.PostDraftOfferBodyModel(
+        body = offers_serialize.PostDraftOfferBodyModel(
             name="A pretty good offer",
             subcategoryId=subcategories.SEANCE_CINE.id,
             venueId=venue.id,
@@ -1189,7 +1189,7 @@ class CreateDraftOfferTest:
 
     def test_cannot_create_draft_offer_with_ean_in_name(self):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.PostDraftOfferBodyModel(
+        body = offers_serialize.PostDraftOfferBodyModel(
             name="A pretty good offer 4759217254634",
             subcategoryId=subcategories.SEANCE_CINE.id,
             venueId=venue.id,
@@ -1221,7 +1221,7 @@ class CreateDraftOfferTest:
             venue=venue,
         )
 
-        body = offers_schemas.PostDraftOfferBodyModel(
+        body = offers_serialize.PostDraftOfferBodyModel(
             name="A pretty good offer",
             subcategoryId=subcategories.SEANCE_CINE.id,
             venueId=venue.id,
@@ -1235,7 +1235,7 @@ class CreateDraftOfferTest:
     def test_create_draft_offer_with_withrawal_details_from_venue(self):
         venue = offerers_factories.VenueFactory(withdrawalDetails="Details from my venue")
 
-        body = offers_schemas.PostDraftOfferBodyModel(
+        body = offers_serialize.PostDraftOfferBodyModel(
             name="A pretty good offer",
             subcategoryId=subcategories.SEANCE_CINE.id,
             venueId=venue.id,
@@ -1246,7 +1246,7 @@ class CreateDraftOfferTest:
 
     def test_cannot_create_activation_offer(self):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.PostDraftOfferBodyModel(
+        body = offers_serialize.PostDraftOfferBodyModel(
             name="An offer he can't refuse",
             subcategoryId=subcategories.ACTIVATION_EVENT.id,
             venueId=venue.id,
@@ -1259,7 +1259,7 @@ class CreateDraftOfferTest:
 
     def test_cannot_create_offer_when_invalid_subcategory(self):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.PostDraftOfferBodyModel(
+        body = offers_serialize.PostDraftOfferBodyModel(
             name="An offer he can't refuse",
             subcategoryId="TOTO",
             venueId=venue.id,
@@ -1278,7 +1278,7 @@ class UpdateDraftOfferTest:
             subcategoryId=subcategories.ESCAPE_GAME.id,
             description="description",
         )
-        body = offers_schemas.PatchDraftOfferBodyModel(
+        body = offers_serialize.PatchDraftOfferBodyModel(
             name="New name",
             description="New description",
         )
@@ -1294,7 +1294,7 @@ class UpdateDraftOfferTest:
             subcategoryId=subcategories.ESCAPE_GAME.id,
             description="description",
         )
-        body = offers_schemas.PatchDraftOfferBodyModel(
+        body = offers_serialize.PatchDraftOfferBodyModel(
             name="New name 4759217254634",
             description="New description",
         )
@@ -1316,7 +1316,7 @@ class CreateOfferTest:
         venue = offerers_factories.VenueFactory(isVirtual=True, offererAddress=None, siret=None)
         offerer_address = offerers_factories.OffererAddressFactory(offerer=venue.managingOfferer)
 
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="A pretty good offer",
             externalTicketOfficeUrl="http://example.net",
             audioDisabilityCompliant=True,
@@ -1335,7 +1335,7 @@ class CreateOfferTest:
         venue = offerers_factories.VenueFactory()
         offerer_address = offerers_factories.OffererAddressFactory(offerer=venue.managingOfferer)
 
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="A pretty good offer",
             subcategoryId=subcategories.SEANCE_CINE.id,
             externalTicketOfficeUrl="http://example.net",
@@ -1375,7 +1375,7 @@ class CreateOfferTest:
         venue = offerers_factories.VenueFactory()
         provider = providers_factories.PublicApiProviderFactory()
 
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="A pretty good offer",
             subcategoryId=subcategories.SEANCE_CINE.id,
             audioDisabilityCompliant=True,
@@ -1396,7 +1396,7 @@ class CreateOfferTest:
 
     def test_cannot_create_activation_offer(self):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="An offer he can't refuse",
             subcategoryId=subcategories.ACTIVATION_EVENT.id,
             audioDisabilityCompliant=True,
@@ -1413,7 +1413,7 @@ class CreateOfferTest:
 
     def test_cannot_create_offer_when_invalid_subcategory(self):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="An offer he can't refuse",
             subcategoryId="TOTO",
             audioDisabilityCompliant=True,
@@ -1428,7 +1428,7 @@ class CreateOfferTest:
 
     def test_cannot_create_offer_with_ean_in_name(self):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="An offer he can't refuse - 4759217254634",
             subcategoryId=subcategories.SEANCE_CINE.id,
             audioDisabilityCompliant=True,
@@ -1458,7 +1458,7 @@ class CreateOfferTest:
     )
     def test_raise_error_if_url_not_coherent_with_subcategory(self, url, subcategory_id, expected_error):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="An offer he can't refuse",
             subcategoryId=subcategory_id,
             url=url,
@@ -1474,7 +1474,7 @@ class CreateOfferTest:
 
     def test_raise_error_if_extra_data_mandatory_fields_not_provided(self):
         venue = offerers_factories.VenueFactory()
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="A pretty good offer",
             subcategoryId=subcategories.CONCERT.id,
             audioDisabilityCompliant=True,
@@ -1503,7 +1503,7 @@ class CreateOfferTest:
             venue=venue,
             idAtProvider=id_at_provider,
         )
-        body = offers_schemas.CreateOffer(
+        body = offers_serialize.CreateOffer(
             name="A pretty good offer",
             subcategoryId=subcategories.SEANCE_CINE.id,
             externalTicketOfficeUrl="http://example.net",
@@ -1531,7 +1531,7 @@ class UpdateOfferTest:
             isDuo=False, bookingEmail="old@example.com", subcategoryId=subcategories.ESCAPE_GAME.id
         )
 
-        body = offers_schemas.UpdateOffer(isDuo=True, bookingEmail="new@example.com")
+        body = offers_serialize.UpdateOffer(isDuo=True, bookingEmail="new@example.com")
         with caplog.at_level(logging.DEBUG):
             offer = api.update_offer(offer, body)
         db.session.flush()
@@ -1560,7 +1560,7 @@ class UpdateOfferTest:
 
     def test_update_extra_data_should_raise_error_when_mandatory_field_not_provided(self):
         offer = factories.OfferFactory(subcategoryId=subcategories.SPECTACLE_REPRESENTATION.id)
-        body = offers_schemas.UpdateOffer(extraData={"author": "Asimov"})
+        body = offers_serialize.UpdateOffer(extraData={"author": "Asimov"})
         with pytest.raises(api_errors.ApiErrors) as error:
             api.update_offer(offer, body)
 
@@ -1573,7 +1573,7 @@ class UpdateOfferTest:
         offer = factories.OfferFactory(
             subcategoryId=subcategories.SPECTACLE_REPRESENTATION.id, extraData={"showType": 200}
         )
-        body = offers_schemas.UpdateOffer(extraData=None)
+        body = offers_serialize.UpdateOffer(extraData=None)
         with pytest.raises(api_errors.ApiErrors) as error:
             api.update_offer(offer, body)
         assert error.value.errors == {
@@ -1588,7 +1588,7 @@ class UpdateOfferTest:
             extraData={"gtl_id": "02000000"},
             subcategoryId=subcategories.SUPPORT_PHYSIQUE_MUSIQUE_CD.id,
         )
-        body = offers_schemas.UpdateOffer(name="New name", description="new Description")
+        body = offers_serialize.UpdateOffer(name="New name", description="new Description")
         offer = api.update_offer(offer, body)
         db.session.flush()
 
@@ -1596,8 +1596,8 @@ class UpdateOfferTest:
         assert offer.description == "new Description"
 
     def test_cannot_update_with_name_too_long(self):
-        offer = factories.OfferFactory(name="Old name", ean="1234567890124")
-        body = offers_schemas.UpdateOffer(name="Luftballons" * 99)
+        offer = factories.OfferFactory(name="Old name", extraData={"ean": "1234567890124"})
+        body = offers_serialize.UpdateOffer(name="Luftballons" * 99)
         with pytest.raises(api_errors.ApiErrors) as error:
             api.update_offer(offer, body)
 
@@ -1605,8 +1605,8 @@ class UpdateOfferTest:
         assert db.session.query(models.Offer).one().name == "Old name"
 
     def test_cannot_update_with_name_containing_ean(self):
-        offer = factories.OfferFactory(name="Old name", ean="1234567890124")
-        body = offers_schemas.UpdateOffer(name="Luftballons 1234567890124")
+        offer = factories.OfferFactory(name="Old name", extraData={"ean": "1234567890124"})
+        body = offers_serialize.UpdateOffer(name="Luftballons 1234567890124")
         with pytest.raises(exceptions.EanInOfferNameException) as error:
             api.update_offer(offer, body)
 
@@ -1618,7 +1618,7 @@ class UpdateOfferTest:
         offer = factories.OfferFactory(
             lastProvider=provider, name="Old name", subcategoryId=subcategories.SEANCE_CINE.id
         )
-        body = offers_schemas.UpdateOffer(name="Old name", isDuo=True)
+        body = offers_serialize.UpdateOffer(name="Old name", isDuo=True)
         api.update_offer(offer, body)
 
         offer = db.session.query(models.Offer).one()
@@ -1630,7 +1630,7 @@ class UpdateOfferTest:
         offer = factories.OfferFactory(
             lastProvider=provider, durationMinutes=90, subcategoryId=subcategories.SEANCE_CINE.id
         )
-        body = offers_schemas.UpdateOffer(durationMinutes=120, isDuo=True)
+        body = offers_serialize.UpdateOffer(durationMinutes=120, isDuo=True)
         with pytest.raises(api_errors.ApiErrors) as error:
             api.update_offer(offer, body)
 
@@ -1647,7 +1647,7 @@ class UpdateOfferTest:
             name="Old name",
             ean="1234567890124",
         )
-        body = offers_schemas.UpdateOffer(externalTicketOfficeUrl="https://example.com")
+        body = offers_serialize.UpdateOffer(externalTicketOfficeUrl="https://example.com")
         api.update_offer(offer, body)
 
         offer = db.session.query(models.Offer).one()
@@ -1664,7 +1664,7 @@ class UpdateOfferTest:
             motorDisabilityCompliant=False,
             mentalDisabilityCompliant=True,
         )
-        body = offers_schemas.UpdateOffer(
+        body = offers_serialize.UpdateOffer(
             name="Old name",
             audioDisabilityCompliant=False,
             visualDisabilityCompliant=True,
@@ -1697,7 +1697,7 @@ class UpdateOfferTest:
             motorDisabilityCompliant=False,
             mentalDisabilityCompliant=True,
         )
-        body = offers_schemas.UpdateOffer(
+        body = offers_serialize.UpdateOffer(
             name="Old name",
             audioDisabilityCompliant=False,
             visualDisabilityCompliant=True,
@@ -1722,7 +1722,7 @@ class UpdateOfferTest:
             audioDisabilityCompliant=True,
             subcategoryId=subcategories.SEANCE_CINE.id,
         )
-        body = offers_schemas.UpdateOffer(
+        body = offers_serialize.UpdateOffer(
             durationMinutes=120,
             isDuo=True,
             audioDisabilityCompliant=False,
@@ -1741,7 +1741,7 @@ class UpdateOfferTest:
 
     def test_update_non_approved_offer_fails(self):
         pending_offer = factories.OfferFactory(name="Soliloquy", validation=models.OfferValidationStatus.PENDING)
-        body = offers_schemas.UpdateOffer(name="Monologue")
+        body = offers_serialize.UpdateOffer(name="Monologue")
         with pytest.raises(exceptions.RejectedOrPendingOfferNotEditable) as error:
             api.update_offer(pending_offer, body)
 
@@ -1760,7 +1760,7 @@ class UpdateOfferTest:
             name="Offer linked to a provider",
             ean="1234567890124",
         )
-        body = offers_schemas.UpdateOffer(idAtProvider="some_id_at_provider")
+        body = offers_serialize.UpdateOffer(idAtProvider="some_id_at_provider")
         api.update_offer(offer, body)
 
         offer = db.session.query(models.Offer).one()
@@ -1776,7 +1776,7 @@ class UpdateOfferTest:
             name="Offer linked to a provider",
             ean="1234567890124",
         )
-        body = offers_schemas.UpdateOffer(ean="1234567890125")
+        body = offers_serialize.UpdateOffer(ean="1234567890125")
         api.update_offer(offer, body)
 
         offer = db.session.query(models.Offer).one()
@@ -1787,7 +1787,7 @@ class UpdateOfferTest:
         offerer = offerers_factories.OffererFactory()
         providers_factories.OffererProviderFactory(offerer=offerer, provider=provider)
         offer = factories.EventOfferFactory(lastProvider=provider, name="Offer linked to a provider", ean=None)
-        body = offers_schemas.UpdateOffer(ean=None)
+        body = offers_serialize.UpdateOffer(ean=None)
         api.update_offer(offer, body)
 
         offer = db.session.query(models.Offer).one()
@@ -1799,7 +1799,7 @@ class UpdateOfferTest:
             name="Offer linked to a provider",
             ean="1234567890124",
         )
-        body = offers_schemas.UpdateOffer(idAtProvider="some_id_at_provider")
+        body = offers_serialize.UpdateOffer(idAtProvider="some_id_at_provider")
         with pytest.raises(exceptions.CannotSetIdAtProviderWithoutAProvider) as error:
             api.update_offer(offer, body)
 
@@ -1828,7 +1828,7 @@ class UpdateOfferTest:
             venue=venue,
         )
 
-        body = offers_schemas.UpdateOffer(idAtProvider=id_at_provider)
+        body = offers_serialize.UpdateOffer(idAtProvider=id_at_provider)
         with pytest.raises(exceptions.IdAtProviderAlreadyTakenByAnotherVenueOffer) as error:
             api.update_offer(offer, body)
 
@@ -1849,7 +1849,7 @@ class UpdateOfferTest:
             url="http://example.com" if isDigital else None,
             subcategoryId=subcategories.VOD.id if isDigital else subcategories.SPECTACLE_REPRESENTATION.id,
         )
-        body = offers_schemas.UpdateOffer(name="New name", offererAddress=offerer_address)
+        body = offers_serialize.UpdateOffer(name="New name", offererAddress=offerer_address)
         if isDigital:
             with pytest.raises(api_errors.ApiErrors) as error:
                 api.update_offer(offer, body)
@@ -1870,7 +1870,7 @@ class UpdateOfferTest:
             offererAddress=venue_oa,
         )
         venue_oa_id = new_venue.offererAddressId
-        body = offers_schemas.UpdateOffer()
+        body = offers_serialize.UpdateOffer()
         assert offer_oa_id != venue_oa_id
 
         updated_offer = api.update_offer(offer, body, venue=new_venue)
@@ -1890,7 +1890,7 @@ class UpdateOfferTest:
             address__city="Lille",
         )
         offer = factories.OfferFactory()
-        body = offers_schemas.UpdateOffer()
+        body = offers_serialize.UpdateOffer()
 
         updated_offer = api.update_offer(offer, body, offerer_address=new_offerer_address)
 
@@ -1909,7 +1909,7 @@ class UpdateOfferTest:
         )
         offer = factories.OfferFactory()
         new_venue = offerers_factories.VenueFactory(managingOfferer=offer.venue.managingOfferer)
-        body = offers_schemas.UpdateOffer()
+        body = offers_serialize.UpdateOffer()
 
         updated_offer = api.update_offer(offer, body, venue=new_venue, offerer_address=new_offerer_address)
 
@@ -2034,7 +2034,7 @@ _FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD = {
 class CreateEventOpeningHoursTest:
     def test_create_event_opening_hours(self):
         offer = factories.EventOfferFactory(subcategoryId=subcategories.FESTIVAL_SPECTACLE.id)
-        body = offers_schemas.CreateEventOpeningHoursModel(**_FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD)
+        body = offers_serialize.CreateEventOpeningHoursModel(**_FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD)
 
         event_opening_hours = api.create_event_opening_hours(body=body, offer=offer)
 
@@ -2044,7 +2044,7 @@ class CreateEventOpeningHoursTest:
 
     def test_create_event_opening_hours_should_raise_because_invalid_category(self):
         offer = factories.EventOfferFactory(subcategoryId=subcategories.ABO_BIBLIOTHEQUE.id)
-        body = offers_schemas.CreateEventOpeningHoursModel(**_FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD)
+        body = offers_serialize.CreateEventOpeningHoursModel(**_FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD)
 
         with pytest.raises(exceptions.OfferEditionBaseException) as error:
             api.create_event_opening_hours(body=body, offer=offer)
@@ -2055,7 +2055,7 @@ class CreateEventOpeningHoursTest:
 
     def test_create_event_opening_hours_should_raise_because_already_has_opening_hours(self):
         offer = factories.EventOpeningHoursFactory().offer
-        body = offers_schemas.CreateEventOpeningHoursModel(**_FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD)
+        body = offers_serialize.CreateEventOpeningHoursModel(**_FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD)
 
         with pytest.raises(exceptions.OfferEditionBaseException) as error:
             api.create_event_opening_hours(body=body, offer=offer)
@@ -2065,7 +2065,7 @@ class CreateEventOpeningHoursTest:
     def test_create_event_opening_hours_should_raise_because_already_has_timestamp_stocks(self):
         offer = factories.EventOfferFactory(subcategoryId=subcategories.FESTIVAL_SPECTACLE.id)
         factories.StockFactory(offer=offer)
-        body = offers_schemas.CreateEventOpeningHoursModel(**_FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD)
+        body = offers_serialize.CreateEventOpeningHoursModel(**_FUNCTIONAL_CREATE_EVENT_OPENING_HOURS_PAYLOAD)
 
         with pytest.raises(exceptions.OfferEditionBaseException) as error:
             api.create_event_opening_hours(body=body, offer=offer)
@@ -2282,7 +2282,7 @@ class UpdateEventOpeningHoursTest:
             ts = timespan_str_to_numrange(time_spans)
             factories.EventWeekDayOpeningHoursFactory(eventOpeningHours=event, timeSpans=ts, weekday=weekday)
 
-        body = offers_schemas.UpdateEventOpeningHoursModel(
+        body = offers_serialize.UpdateEventOpeningHoursModel(
             startDatetime=_base_start_datetime.astimezone(pytz.utc),
             endDatetime=_base_end_datetime.astimezone(pytz.utc),
             openingHours={**base_update_opening_hours_payload(), **new_opening_hours},
